Refactor in preparation for running the spec harness
