+++ lib/item_repository.rb
@@ -44,4 +44,6 @@ class ItemRepository
       price_range.include?(item.unit_price_to_dollars)
     end
   end
+
+  def inspect; end
 end
