+++ lib/merchant.rb
@@ -1,4 +1,4 @@
-require './lib/sales_engine'
+require_relative './sales_engine'
 
 class Merchant
