+++ lib/merchant_repository.rb
@@ -26,9 +26,6 @@ class MerchantRepository
       merchant.name.downcase.include?(name_stub.downcase)
     end
   end
+
+  def inspect; end
 end
-
-
-
-
-
