+++ lib/sales_engine.rb
@@ -1,6 +1,6 @@
-require './lib/merchant_repository'
-require './lib/item_repository'
-require './lib/importer'
+require_relative './merchant_repository'
+require_relative './item_repository'
+require_relative './importer'
 
 class SalesEngine
