+++ test/customer_test.rb
@@ -1,6 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/customer'
+require_relative '../lib/customer'
 
 class CustomerTest < Minitest::Test
