+++ test/invoice_test.rb
@@ -1,6 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/invoice'
+require_relative '../lib/invoice'
 
 class InvoiceTest < Minitest::Test
