+++ test/item_test.rb
@@ -1,6 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/item'
+require_relative '../lib/item'
 
 class ItemTest < Minitest::Test
