+++ test/merchant_test.rb
@@ -1,6 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/merchant'
+require_relative '../lib/merchant'
 
 class MerchantTest < Minitest::Test
