+++ test/sales_engine_test.rb
@@ -1,8 +1,8 @@
 require 'minitest/autorun'
 require 'minitest/pride'
-require './lib/sales_engine'
-require './lib/merchant_repository'
-require './lib/item_repository'
+require_relative '../lib/sales_engine'
+require_relative '../lib/merchant_repository'
+require_relative '../lib/item_repository'
 
 class SalesEngineTest < Minitest::Test
 
@@ -49,23 +49,6 @@ class SalesEngineTest < Minitest::Test
     })
     assert_equal number_of_items_in_file, sales_engine.items.all.length
     assert sales_engine.items.find_by_name("Anello nodo")
-  end
-
-  def test_it_can_import_the_provided_file_for_merchants
-    sales_engine = SalesEngine.from_csv({
-      :merchants => './data/merchants.csv'
-    })
-    assert_equal 475, sales_engine.merchants.all.length
-    assert sales_engine.merchants.all.map(&:name).include?('Woodenpenshop')
-  end
-
-  def test_it_can_import_the_provided_file_for_items
-    skip "for speed!"
-    sales_engine = SalesEngine.from_csv({
-      :items => './data/items.csv'
-    })
-    assert_equal 1367, sales_engine.items.all.length
-    assert sales_engine.items.all.map(&:name).include?('wooden finger protection')
   end
 
   def test_it_can_find_items_by_merchant_id
